Added waitgroups for autoupdate workers to complete before stopping
The wait.UntilWithContext will repeatedly run the function passed to it
till the context also passed to it is done. At which point it will
return. Adding a wait group after the UntilWithContext ensures that
the top level goroutine does not return till all the spawned goroutines
terminate. Also added `HandleCrash()` to log any panics in inner
goroutines.