Added ownerReference to all objects created and managed by CVO
The operator object no has a metadata modifier which injects an owner reference if there is no existing owner references. Also updates the ownerReference if it has been changed due to recreation. The ClusterVersion get-or-create method has also been modified so that ClusterVersion is refetched after creation to ensure that the CV UID is present.