pkg/cvo: Respect noProxy
Since 4.2's ea5e3bc86e (Add http transport for cincinnati to enable
proxy, 2019-07-16, #219), the CVO has been consuming httpsProxy, but
not httpProxy or noProxy [1].  This commit fixes that.  I'm handing
off the logic that mixes those three together to pick the proxy URI to
the httpproxy library, which is also what the Go standard library uses
for this purpose.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1978749