Log resource diffs on update only in reconcile mode
Logging of resource diffs was introduced to detect hotlooping in the
CVO[1]. But this increases the verbosity of logging during updates when
the resources have changed which is not useful. Hence logging of the
diffs is more useful in the reconcile mode. Also the diff is generated
with ObjectDiff[2] which is deprecated. The ObjectDiff function now just
uses cmp.Diff[3] internally and all instances of ObjectDiff have been replaced with
cmp.Diff.

[1] - https://github.com/openshift/cluster-version-operator/pull/561
[2] - https://github.com/kubernetes/apimachinery/blob/235edae7dd90601011bbe3bcd6f84f7dc857b034/pkg/util/diff/diff.go#L57
[3] - https://pkg.go.dev/github.com/google/go-cmp/cmp#Diff