Fail fast when applying Custom Resources with no Definition
The CVO attempts to retry creation of resources when there are failures.
But this approach prevents CVO from advancing when the resource defition
for a resource is missing. The CVO should instead no rety and proceed
with the creation of subsequent resources so that the definition could
be applied later.

Bug 1927168: The deleted crd/servicemonitors.monitoring.coreos.com is not recreated quickly