pkg/cvo: Drop the explicit 'upstream' from our replacement ClusterVersion
Hopefully we never actually have to stuff a CVO-generated
ClusterVersion into the cluster; it's just for recovery after admins
accidentally delete their existing ClusterVersion.  But if we ever do
hit this code, we want to push it without spec.upstream, to allow
later ClusterVersion-consuming code to say "ah, user doesn't care
which upstream I use, so I'll use the best default I'm aware of".
This effectively pushes default choice from the CVO that creates the
ClusterVersion out to the CVO that consumes the ClusterVersion, and
that later CVO is almost certainly more current on which default is
best.

Similar to openshift/installer@c9095b3451
(data/manifests/bootkube/cvo-overrides: Drop the explicit upstream,
2020-08-28, openshift/installer#4112).