pkg/cvo/availableupdates: Expose upstream URI in RetrievedUpdates=True
So folks can easily see where we're getting the data from, even if
spec.upstream is unset.  I'm not including the proxy information,
because that's all calculated in other layers, and it may contain
sensitive information like proxy-connection credentials.  We don't
have to worry about sensitive information in the upstream URI, because
our default value is in public code, and any sensitive information
from spec.upstream is already in the ClusterVersion object where we're
putting this message.