pkg/cvo/upgradeable: Enable admin-ack logic
519b466793 (Bug 1978376: Add admin ack Upgradeable condition gate,
2021-07-27, #633) had these commented out, because 4.9 has no built-in
acks.  But with the code commented out, it's hard to verify that the
logic works in 4.9 before backporting to 4.8 [1].  Enabling these
checks should be a no-op outside of verification, because admins are
unlikely to inject additional keys in the openshift-config-managed
namespace's admin-gates ConfigMap.  And it allows us to verify the
logic in 4.9 and cook there with live code before approving the 4.8
backports.  It's also one less thing we might forget before enabling
new admin acks in future versions, like 4.10 or later.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1978376#c19