lib/resourcemerge/core: Remove unrecognized volumes and mounts
Since this package was created in d9f6718de0 (lib: add lib for
applying objects, 2018-08-14, #7), the volume(mount) merge logic has
required manifest entries to exist, but has allowed in-cluster entries
to persist without removal.  That hasn't been a problem until [1]:

1. In 4.3, the autoscaler asked for a ca-cert volume mount, based on
   the cluster-autoscaler-operator-ca config map.
2. In 4.4, the autoscaler dropped those manifest entries [2].
3. In 4.9, the autoscaler asked the CVO to remove the config map [3].

That lead some born-in 4.3 clusters to have crashlooping autoscalers,
because the mount attempts kept failing on the missing config map.

We couldn't think of a plausible reason why cluster admins would want
to inject additional volume mounts in a CVO-managed pod configuration,
so this commit removes that ability and begins clearing away any
volume(mount) configuration that is not present in the reconciling
manifest.  Cluster administrators who do need to add additional mounts
in an emergency are free to use ClusterVersion's spec.overrides to
take control of a particular CVO-managed resource.

This joins a series of similar previous tightenings, including
02bb9ba829 (lib/resourcemerge/core: Clear env and envFrom if unset in
manifest, 2021-04-20, #549) and ca299b8871 (lib/resourcemerge: remove
ports which are no longer required, 2020-02-13, #322).

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=2002834
[2]: https://github.com/openshift/cluster-autoscaler-operator/commit/f08589d4767037636fc0d3e78b40d61f9eed8104#diff-547486373183980619528df695869ed32b80c18383bc16b57a5ee931bf0edd39L89
[3]: https://github.com/openshift/cluster-autoscaler-operator/commit/9a7b3be569980316235990fe1feb106f0155e4b2#diff-d0cf785e044c611986a4d9bdd65bb373c86f9eb1c97bd3f105062184342a872dR4