Bug 1990635: Removing ValidateClusterVersion from the availableUpdatesSync method
As availableUpdatesSync() does not need  the ValidateClusterVersion()
check.
The purpose of ValidateClusterVersion() is to make sure the version is
valid e.g. user provided a version but it checks if the version is associated
with a pull spec or not. However the objective of availableUpdatesSync()
to find if there are versions available to upgrade. These two methods
are mutually exclusive.

If we check the validity of the version before checking
syncAvailableUpdates() then it does not get to the available versions.

Signed-off-by: Lalatendu Mohanty <lmohanty@redhat.com>