pkg/cvo/upgradeable: Include messages for multiple-reason Upgradeable=False
Because:

  Upgradeable=False

    Reason: MultipleReasons
    Message: Cluster cannot be upgraded between minor versions for multiple reasons: AdminAckRequired,IncompatibleOperatorsInstalled

doesn't include all the useful information needed to resolve those
issues.  This pivots to using the same approach we use today when
aggregating multiple Upgradeable=False ClusterOperators.

Also shift the patch-blocking overrides check in front of the other
minor blocking checks.  It's more important (because it blocks more
targets), and it also will not included a nested sub-list (which might
be confusing, because today I'm not adding indents or any such markers
to make it easy to distinguish levels in a nested list hierarchy).

This cherry-picks 0c742e5666 (pkg/cvo/upgradeable: Include messages
for multiple-reason Upgradeable=False, 2021-10-07, #671) back to 4.8,
with a minor context change because 4.8 doesn't have 0afb8a8658 (Add a
manifest annotation to be used for object deletion, 2020-08-17, #438).