bootstrap/bootstrap-pod: Pivot from --enable-default-cluster-version to --wait-for-cluster-version
We've been using --enable-default-cluster-version to avoid a
cluster-bootstrap vs. cluster-version operator race since d7760cec0f
(pkg/cvo: Drop ClusterVersion defaulting during bootstrap, 2019-08-16, #238).
But [1] is considering passing component configuration to the CVO via
ClusterVersion that would affect the manifests getting applied.  So
this commit pivots to a new option that blocks the whole sync cycle on
the existence of a ClusterVersion resource.  That way, we know we have
up-to-date data before reconciling any manifests (although it might
slow down the bootstrapping process a bit).

[1]: https://github.com/openshift/enhancements/pull/922