lib/resourcemerge/core: Merge volumeMounts by mountPath
We had been merging by name since ensureVolumeMounts landed in
83faa6e716 (lib/resourcemerge/core: Remove unrecognized volumes and
mounts, 2021-09-14, #654).  But as pointed out in [1], a single volume
may be mounted at multiple paths.  With this commit, I'm pivoting to
merge by mountPath, which is the patchMergeKey [2] (and it makes sense
that you wouldn't have multiple volumes mounted at the same path).

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=2026560
[2]: https://github.com/kubernetes/api/blob/1d6faf224f146dd002553f55cd9fcaaaa0dc00cb/core/v1/types.go#L2367