pkg/start: Log and continue when we fail to retrieve the feature gate
From the enhancement [1]:

  During bootstrapping, the CVO will assume no feature sets are
  enabled until it can successfully retrieve
  `featuregates.config.openshift.io` from the Kubernetes API server.

So this softens the error from 18dd189635 (simplify includeTechPreview
flag to be a static bool, 2021-11-22, #694) to be log-and-continue.
If it turns out that we actually were TechPreviewNoUpgrade, the
TechPreviewChangeStopper controller will eventually succeed in pulling
the feature-gate, notice, and ask the CVO to shut down.  In the
meantime, the CVO will have been able to get a head start on
reconciling the vast majority of manifests which are not tech-preview.

[1]: https://github.com/openshift/enhancements/blame/f74ffe7776f40dbd096b9ca10c27ee7a0a579e58/enhancements/update/cvo-techpreview-manifests.md#L70-L71