pkg/payload/precondition: Drop ClusterVersion argument
The argument landed in b72e843974 (Bug 1822844: Block z level upgrades
if ClusterVersionOverridesSet set, 2020-04-30, #364) for use by
Upgradeable.Run.  But even then, that method opens by retrieving a
(possibly cached) ClusterVersion resource from the configured lister,
so there's no need to pass the explicit argument.  We should save
explicit inputs for things that need to be passed in from memory at
call-time, and not use them for information that can be retrieved from
precondition-creation-time callbacks.  And even for things that need
to come from memory at call time, we should be using ReleaseContext so
we can add and remove properties without having to touch function
signatures for precondition implementations that don't care about the
properties we're touching.

While I'm touching the Run call site, I replaced a context.TODO with a
context.Background.  As pointed out in the docs [1], Background is
prefered for tests.

[1]: https://pkg.go.dev/context#Background