pkg/cvo/availableupdates: Acount for default upstream in recent-change throttling
Since [1], clusters born in 4.7 or later default to not having an
explicit spec.upstream, so they rely on the CVO's internal default.
However, in that case, u.Upstream will be an empty string, while by
this point in syncAvailableUpdates, upstream will have been set to the
default value.  This commit splits up the "do we really need to
check?" logic into a number of distinct cases, and gives them more
specific logging, to make it easier to understand and confirm the
desired behavior:

a. If we have no cached data, we need to pull a graph.
b. If it's been over minimumUpdateCheckInterval since our last check,
   we need to pull a graph.  Even if nothing has changed on our side,
   our data is sufficiently stale to need a refresh.
c. If the channel has changed, we have different interests, and we
   need to pull a graph to hear what the upstream recommends for this
   new set of interests.
d. If the upstream hasn't changed, because:
   i. The current upstream (explicitly or by default) matches the old
      explicit upstream, or
   ii. The current upstream (explicitly or by default) matches the
       default, and the old upstream was unset.
   then everything's the same on our side, and our cached graph is
   recent, so we don't need to do anything.
e. Otherwise, the upstream has changed, and we need to pull a graph to
   see what our new guide has to suggest.

Cases for upstream:

* A -> A: Handled by d.i.
* A -> B: Handled by e.
* A -> unset (defaulted) or default: Handled by e.
* Unset or default -> A: Handled by e.
* Default -> default: Handled in d.i.
* Default -> unset (defaulted): Handled in d.i.
* Unset -> default: Handled by d.ii, new in this commit, previously
  resulted in an excessive pull.
* Unset -> unset (defaulted): Handled by d.ii, new in this commit,
  previously resulted in an excessive pull.

[1]: https://github.com/openshift/installer/pull/4112