pkg/cvo/metrics: Add label 'verified' to metric 'cluster_version'
Add the label 'verified' to the metric 'cluster_version' to indicate
whether a provided update was properly verified before it was installed.
This information will get pushed up to telemetry, so we can better
analyze the provided information.

Modify the code appropriately to use the value for verified which is
stored in the 'ClusterVersion.Status.History'. To get the value of
'verified' of a current version, the code will search the
'ClusterVersion.Status.History', as to not report the value incorrectly
in a scenario where the specific update hasn't had the time to be added
to the history. If no such update can be found in the history,
verified="" is reported.

Modify the metrics tests to test the implemented changes.