install/0000_90_cluster-version-operator_02_servicemonitor: Alert if current version of cluster is not verified
Add the alert `CurrentVersionNotVerified` to inform the user about the
current version of the cluster not being verified. These alerts will
also be pushed up into the Telemetry to help us analyze what portion of
the clusters are not verified.

Set the alert's severity to `info`. The `warn` severity would be too
unactionable. But there's a value for an info alert providing additional
context when investigating cluster issues. For an extensive explanation
for the severity level, see: https://github.com/openshift/cluster-version-operator/pull/734#discussion_r826778130