lib/resourcebuilder/batch: Stop waiting on Job deadline exceeded
Avoid waiting for a job that that job controller has given up on [1].
With this change, when we have issues like DeadlineExceeded, we will
set 'err != nil' for "things are not good" and (new in this commit)
'done == true' for "waiting longer is useless".  Also split the 'err
!= nil' cases to "log the error and wait longer" and (new in this
commit) "bubble the error up to the caller" cases.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=2062568#c0