pkg/cvo/updatepayload: Event when forcing through a sig-verification failure
For unforced updates, when signature, etc. verification fails,
RetrievePayload returns an error, and we have emitted events for
RetrievePayload errors since 475e71f334 (emit events for each new
payload, 2020-07-21, #411).  However, when the user forces the update,
we log but do not error out on verification failures.  With this
commit, we will also emit a warning event with an error message, which
will make it easier to understand how signature verification failed,
even when we don't have access to the logs of the outgoing
cluster-version operator [1].

We should transition away from the maintenance-mode
github.com/pkg/errors [2], but I'm deferring that to follow-up work.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=2071998#c2
[2]: https://pkg.go.dev/github.com/pkg/errors#readme-roadmap