Do not save desired update on load failures
since this will cause equalSyncWork to continually return
true and no reattempt to load the payload will occur until
the desired update is updated by the user again, e.g.
image change or change to force.

As a result, no attempt is made to recheck precondition
failures which may have been resolved and therefore
result in a successful payload load.

To take the specific
https://bugzilla.redhat.com/show_bug.cgi?id=2072389 bug
failure as an example, because no recheck is made on
the RecentEtcdBackup precondition CVO does not detect
that the Etcd backup has been completed and it's safe
to continue with the update. In addition, as a result
of change
https://github.com/openshift/cluster-version-operator/pull/683,
the etcd operator must check ReleaseAccepted!=true rather
than Failing=true to trigger the start of the backup.