pkg/cvo/updatepayload: Use initContainers instead of shell &&-chains
All of the input strings were internally controlled, so there wasn't a
risk of them containing a shell-sensitive character.  But that safety
was not immediately obvious when reading the code.  By converting to
initContainers, which run sequentially to completion before the next
container is run [1], we have the same effect without involving the
shell.  As a side benefit, we also get clearer status logging showing
exactly which steps have succeeded or failed.

[1]: https://kubernetes.io/docs/concepts/workloads/pods/init-containers/#understanding-init-containers