lib/resourcebuilder: Use actual resource in check*Health calls
I'd accidentally broken this in e8ffccb878 (lib: Add autogeneration
for some resource* functionality, 2020-07-29, #420), when I moved the
health checks out of the Do methods and began passing typedObject into
them.  typedObject is the release manifest, which lacks the status
attributes we want the health checks to cover.

This also catches the generation script up with some past manual
changes:

* 0afb8a8658 (Add a manifest annotation to be used for object
  deletion, 2020-08-17, #438).
* 05e1af7fba (Log resource diffs on update only in reconcile mode,
  2021-07-13, #628).