lib/resourcebuilder: Drop Get from check*Health functions
I'd missed these Gets in 0833bba34d (lib/resourcebuilder: Use actual
resource in check*Health calls, 2022-05-03, #771).  What I'd thought
was happening:

1. Pass typedObject into check*Health.
2. No status on typedObject, so check*Health passed, even if the
   in-cluster resource was sad.

What was actually happening:

1. Pass typedObject into check*Health.
2. check*Health calls Get to see what's going on in the cluster.
3. Health check appropriately checks the health of the in-cluster resource.

However, now that 0833bba34d is passing in the just-retrieved
in-cluster resource, the check*Health Get call is an API call that we
don't need to make.  Dropping it here saves us and the API server a
small amount of CPU cycles and network bandwidth.