pkg/clusterconditions/promql: Cap PromQL queries at 5 minutes
In some clusters, these PromQL queries can hang for hours, possibly
forever [1].  I think we have a 30s default KeepAlive timeout [2], but
apparently there's enough socket traffic to keep from tripping that.
This adds a 5m cap to the PromQL calls, although I'm not particularly
attached to that particular number.  We can always raise it if we
start seeing timeouts in Insights for queries where taking that long
seems reasonable.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=2109374#c12
[2]: https://pkg.go.dev/github.com/prometheus/client_golang/api#pkg-variables