pkg/cvo/upgradeable.go: Sync Upgradeable status of the CVO more often
Synchronize the CVO's Upgradeable status more often for a given period
of time when and after the precondition checks start to fail. We don't
want to check this more frequently forever in the case of the
precondition checks failing because of a bigger problem that wasn't
quickly solved at the start of the upgrade by itself.

A precondition check can be failing for at least
`optr.minimumUpdateCheckInterval` time because of `Upgradeable==false`.
The `Upgradeable==false` could have been potentially already resolved
(operator reporting `Upgradeable==false` just momentarily) during the
duration resulting in necessary waiting for up to
`optr.minimumUpdateCheckInterval` for the next synchronization.
Synchronize the upgradeable status again in case of failing
precondition checks to speed up initializing an upgrade stuck on
a precondition check that potentially has been solved. We don't want to
check this forever in case of precondition checks failing for a long
time due to a bigger problem.

This commit is part of a fix for the bug [1].

The bug was caused by slow syncing of the CVO's Upgradeable status when
a precondition check fails and less frequent running of the precondition
checks. The frequency of precondition checks was fixed by [2] fixing the
Etcd backup halting an upgrade for a prolonged time [3]. The problem of
the `Upgradeable==false` thanks to the
`ErrorCheckingOperatorCompatibility` caused by the OLM operator [1] was
fixed by the [3]. However, the main root cause of the
`ErrorCheckingOperatorCompatibility` error probably still remained.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2006611

[2] https://github.com/openshift/cluster-version-operator/pull/683

[3] https://bugzilla.redhat.com/show_bug.cgi?id=2072348

[4] https://github.com/openshift/cluster-version-operator/pull/766