pkg/cvo: add test utility functions
that allows checks of an entire structure or one of it's sub-structures
(using generics). Also added is the ability to check for a single
status, e.g. the "final" status, by specifying a condition to indicate
some state such as finality, e.g. Completed==3. There are also some
functions that allow a test to wait for a particular condition,
e.g. payload load status is "PayloadLoaded", before continuing.