resourceapply: improve diff logging for deployments
Previously we logged the diff between `existing` and `required`, but
`existing` already contains the result of the merge and therefore all
relevant bits from `required` are already identical and will never be
logged, so the logged diff is pretty much always just misleading.

Instead, save the original existing structure and compare it to the
result of the merge. Also, correctly handle the empty diff shown on
coding errors in `EnsureDeployment` when there is no relevant difference
but somehow `modified` still was set to `true`.