resourcemerge: treat empty `PodSecurityContext` as nil
For some reason the API/client sometimes returns a pointer to an empty
`PodSecurityContext` struct instead of a `nil` pointer. When our desired
state is `nil`, this leads to a NOP update and a hotloop. If the desired
state is `nil`, do not modify the existing state if it is either `nil`
or equal to an empty struct.