Poll for 30s on network errors when getting featuresets
The client-go code retries a subset of network errors on GET for 30s,
but we saw occurrences of other short disruptions, like DNS ones, that
make us abort and restart unnecessarily soon.

Make CVO retry all errors for 25s and only abort when we do not succeed
in this time frame. This helps CVO survive short disruptions on startup,
leading to less noise, mostly during installation.