OCPBUGS-2592: Fix Deployment hotloops after incorrect SeccompProfile comparison (#855)
* resourceapply: improve diff logging for deployments

Previously we logged the diff between `existing` and `required`, but
`existing` already contains the result of the merge and therefore all
relevant bits from `required` are already identical and will never be
logged, so the logged diff is pretty much always just misleading.

Instead, save the original existing structure and compare it to the
result of the merge. Also, correctly handle the empty diff shown on
coding errors in `EnsureDeployment` when there is no relevant difference
but somehow `modified` still was set to `true`.

* resourcemerge: fix `SeccompProfile` comparison

`ensureSeccompProfile` was comparing a pointer to a struct and hence
always considered the two `SeccompProfile`s to be different, set
`modified` to true and triggered the update path for possibly unchanged
`Deployments`

Resolves: OCPBUGS-2592

* resourcemerge: treat empty `PodSecurityContext` as nil

For some reason the API/client sometimes returns a pointer to an empty
`PodSecurityContext` struct instead of a `nil` pointer. When our desired
state is `nil`, this leads to a NOP update and a hotloop. If the desired
state is `nil`, do not modify the existing state if it is either `nil`
or equal to an empty struct.