lib: Add /apis/operators.coreos.com/v1/operatorgroups handling
Add /apis/operators.coreos.com/v1/operatorgroups handling that enables
a custom way of processing manifests representing the OperatorGroup
resource. Currently, the OperatorGroup resources are reconciled as
generic resources. This doesn't provide defaulting of the fields.

This commit will provide more precise processing of the OperatorGroup
resources and will provide defaulting of the resources, which will fix
the current hotlooping of the CVO.

Currently, the CVO hotloops on the OperatorGroup resources due to the
missing defaulting of the resource (Bug 2089093 [1]).

The [2] commit introduced the UpgradeStrategy field to the OperatorGroup
type. This field has a default value of "Default". The CVO currently
doesn't know of this default value, and thus the CVO hotloops when
handling the OperatorGroup resource as a generic resource because some
manifests don't specify the value of the field and rely on
defaulting [3].

This commit will provide a more precise way of processing the
OperatorGroup type and will also fix the mentioned bug by providing
the defaulting for the resource.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2089093
[2] https://github.com/openshift/operator-framework-olm/commit/d895574f25098093ccde2108c7cb8e288972f1d7
[3] https://github.com/openshift/operator-framework-olm/blob/ab5ecfd7f235142f191d91d103afabccdfd6ff04/manifests/0000_50_olm_13-operatorgroup-default.yaml