pkg/cvo/updatepayload: Restructure desired-image lookup
The previous logic ignored 'version' if 'image' was specified.  This
allowed admins to request incompatible (image, version) tuples, and
the cluster-version operator would silently accept their 'image' and
not warn the admin that the actual version diverged from their
request.

When desiredUpdates has any opinions, the new logic shuffles to create
a set of targets, including:

* The current release.
* Everything in availableUpdates.
* If desiredUpdate.image is not empty, and also not a digest-match for
  the current release or any availableUpdates, also a new entry with
  information from the desiredUpdate request.

That set of targets is filtered to find any that match all populated
desiredUpdate properties.  So if only desiredUpdate.image is set, we
look for any targets that match that image digest.  If only version is
set, we look for any targets that match that version.  If both image
and version are set, we look for any targets that match both that
image and that image digest.

In the event that we find no matching targets, or that we find
multiple matching targets, we complain, and stick with our current
release.

For now, to preserve backwards compatibility with the previous
implementation, we warn, but still accept the image target, if an
incompatible (image, version) tuple is requested.