lib/resourcemerge: Fix removing caBundle field of CRDs when "inject-cabundle=true"
This commit will fix CVO's behavior of removing the `caBundle` field of
CRDs when reconciling manifests as described in the bug [1].

Currently, the users can omit the `caBundle` field in the CRDs and set
the `service.beta.openshift.io/inject-cabundle=true` annotation so that
the given `caBundle` field is managed by the service-ca controller
(this functionality is not limited only to CRDs) [2].

However, the CVO is not aware of this feature. CVO, when reconciling
such CRD, sees an empty `caBundle` field in the given release manifest,
and CVO then removes any value in the existing applied CRD so it matches
the release manifest semantically.

This commit makes the CVO look into the existing resource and check for
the `service.beta.openshift.io/inject-cabundle=true` annotation. If it's
set then the service-ca controller is responsible for the field, and CVO
will ignore any differences between the existing `caBundle` field and
the `caBundle` field in the given release manifest of the CRD.

This bug also caused the CVO to hotloop on CRDs as mentioned in the bug [3].

Even though the `.Spec.Conversion.Webhook.ClientConfig` is required to
be set if `.Spec.Conversion.Strategy` is set to "Webhook", check for its
existence. A developer might have written an incorrect YAML manifest
which could result in CVO looking at `nil` address, resulting in a
panic. Check for nil pointers in the `existing` instance as the instance
could have been rewritten entirely by the user.

Add a new file for the new constant `injectCABundleAnnotation` since the
constant can be used across multiple files in the `resourcemerge`
package. The annotation can be used in multiple groups for different
kinds (CRD, API service, webhook configurations) [2]. Don't export the
constant because its use is only intended for the `resourcemerge`
package.

[1] https://issues.redhat.com/browse/OCPBUGS-702
[2] https://docs.openshift.com/container-platform/4.11/security/certificates/service-serving-certificate.html
[3] https://bugzilla.redhat.com/show_bug.cgi?id=2089097